requirements.txt
attrs==18.1.0
click==6.7
peewee==3.6.4
PyYAML==3.13
